Add Provision Arguments to ToxParser
