+++ docs/changelog/3190.bugfix.rst
@@ -1,0 +1,1 @@
+Add provision arguments to ToxParser to fix crash when provisioning new tox environment without list-dependencies by :user:`seyidaniels`
