+++ src/tox/session/cmd/run/common.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import logging
 import os
-import random
-import sys
 import time
 from argparse import Action, ArgumentError, ArgumentParser, Namespace
 from concurrent.futures import CancelledError, Future, ThreadPoolExecutor, as_completed
@@ -21,7 +19,6 @@ from tox.execute import Outcome
 from tox.journal import write_journal
 from tox.session.cmd.run.single import ToxEnvRunResult, run_one
 from tox.tox_env.runner import RunToxEnv
-from tox.util.ci import is_ci
 from tox.util.graph import stable_topological_sort
 from tox.util.spinner import MISS_DURATION, Spinner
 
@@ -62,17 +59,8 @@ class InstallPackageAction(Action):
         setattr(namespace, self.dest, path)
 
 
-def env_run_create_flags(parser: ArgumentParser, mode: str) -> None:  # noqa: C901
+def env_run_create_flags(parser: ArgumentParser, mode: str) -> None:
     # mode can be one of: run, run-parallel, legacy, devenv, config
-    if mode not in {"config", "depends"}:
-        parser.add_argument(
-            "--result-json",
-            dest="result_json",
-            metavar="path",
-            of_type=Path,
-            default=None,
-            help="write a JSON file with detailed information about all commands and results involved",
-        )
     if mode not in {"devenv", "depends"}:
         parser.add_argument(
             "-s",
@@ -115,69 +103,11 @@ def env_run_create_flags(parser: ArgumentParser, mode: str) -> None:  # noqa: C9
             dest="develop",
         )
     if mode != "depends":
-
-        class SeedAction(Action):
-            def __call__(
-                self,
-                parser: ArgumentParser,  # noqa: ARG002
-                namespace: Namespace,
-                values: str | Sequence[Any] | None,
-                option_string: str | None = None,  # noqa: ARG002
-            ) -> None:
-                if values == "notset":
-                    result = None
-                else:
-                    try:
-                        result = int(cast(str, values))
-                        if result <= 0:
-                            msg = "must be greater than zero"
-                            raise ValueError(msg)  # noqa: TRY301
-                    except ValueError as exc:
-                        raise ArgumentError(self, str(exc)) from exc
-                setattr(namespace, self.dest, result)
-
-        if os.environ.get("PYTHONHASHSEED", "random") != "random":
-            hashseed_default = int(os.environ["PYTHONHASHSEED"])
-        else:
-            hashseed_default = random.randint(1, 1024 if sys.platform == "win32" else 4294967295)  # noqa: S311
-
-        parser.add_argument(
-            "--hashseed",
-            metavar="SEED",
-            help="set PYTHONHASHSEED to SEED before running commands. Defaults to a random integer in the range "
-            "[1, 4294967295] ([1, 1024] on Windows). Passing 'notset' suppresses this behavior.",
-            action=SeedAction,
-            of_type=Optional[int],
-            default=hashseed_default,
-            dest="hash_seed",
-        )
-    parser.add_argument(
-        "--discover",
-        dest="discover",
-        nargs="+",
-        metavar="path",
-        help="for Python discovery first try the Python executables under these paths",
-        default=[],
-    )
-    if mode != "depends":
         parser.add_argument(
             "--no-recreate-pkg",
             dest="no_recreate_pkg",
             help="if recreate is set do not recreate packaging tox environment(s)",
             action="store_true",
-        )
-        list_deps = parser.add_mutually_exclusive_group()
-        list_deps.add_argument(
-            "--list-dependencies",
-            action="store_true",
-            default=is_ci(),
-            help="list the dependencies installed during environment setup",
-        )
-        list_deps.add_argument(
-            "--no-list-dependencies",
-            action="store_false",
-            dest="list_dependencies",
-            help="never list the dependencies installed during environment setup",
         )
     if mode not in {"devenv", "config", "depends"}:
         parser.add_argument(
